convert in dynamic component
